refactor for broader use
